Move IClientCapabilitiesService down to workspaces, and update namespace
