+++ src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/Protocol/IClientCapabilitiesService.cs
@@ -3,7 +3,7 @@
 
 using Microsoft.VisualStudio.LanguageServer.Protocol;
 
-namespace Microsoft.AspNetCore.Razor.LanguageServer;
+namespace Microsoft.CodeAnalysis.Razor.Workspaces.Protocol;
 
 internal interface IClientCapabilitiesService
 {
